Enhance map zoom and search functionality
The map zoom now has more granular control with the addition of zoomDelta, zoomSnap, and wheelPxPerZoomLevel options. The search functionality has been refined to only return results with an 'osm_key' of 'place' and a name length of at least 3 characters. This helps to ensure that only relevant and meaningful results are returned to the user.