feat: add validation for search query length
Implemented a feature to check the length of the search query in the map search function. Now, the function will only proceed if the query length is at least 3 characters long. If not, it will display a message asking the user to enter at least 3 letters. Also, restructured the code for readability.