feat: Add search functionality to map UI
This update adds an event listener to the search input field. When a user starts to type in the search field, the favorite places container will be hidden. If the search field is cleared, the favorite places container will be displayed again. This enhances the user experience by reducing clutter and focusing on the search results when a search is initiated.